backend/requirements.txt
# A股打板提示工具 - Python依赖
# 后端框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 定时任务
apscheduler>=3.10.0

# 配置管理
pyyaml>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# WebSocket
websockets>=12.0

# 工具库
python-dateutil>=2.8.0
pytz>=2024.1
loguru>=0.7.0

# 数据源 (可选，不安装则使用模拟数据)
# adata>=2.9.0
